Sets up NTBH theme
This is the first theme to be set up, so there's a bit of uncertainty here. I took a very old website I worked on a few years ago, and stripped out all the superfluous HTML and all the styles. This update is a work in progress, with some scaffolding done, but no real styling yet.